Completions - extract and test non-sublime logic (#580)
* Move completion parsing/formatting to core

This allows unit testing this functionality, more scenarios (by passing in settings)
Also removes a guard for textedits on the wrong row, and pre-calculates where a textedit based completion should start from.

* Clean up old formatting tests

* Add tests for parsing and preferring label over filtertext
