+++ plugin/completion.py
@@ -11,14 +11,13 @@ from .core.protocol import Request
 from .core.events import global_events
 from .core.settings import settings, client_configs
 from .core.logging import debug
-from .core.protocol import CompletionItemKind, Range
+from .core.completion import parse_completion_response
 from .core.registry import session_for_view, client_for_view
 from .core.configurations import is_supported_syntax
 from .core.documents import get_document_position
 from .core.sessions import Session
 
 NO_COMPLETION_SCOPES = 'comment, string'
-completion_item_kind_names = {v: k for k, v in CompletionItemKind.__dict__.items()}
 
 
 class CompletionState(object):
@@ -169,58 +168,12 @@ class CompletionHandler(sublime_plugin.ViewEventListener):
                     self.handle_error)
                 self.state = CompletionState.REQUESTING
 
-    def format_completion(self, item: dict) -> 'Tuple[str, str]':
-        # Sublime handles snippets automatically, so we don't have to care about insertTextFormat.
-        if settings.prefer_label_over_filter_text:
-            trigger = item["label"]
-        else:
-            trigger = item.get("filterText", item["label"])
-        # choose hint based on availability and user preference
-        hint = None
-        if settings.completion_hint_type == "auto":
-            hint = item.get("detail")
-            if not hint:
-                kind = item.get("kind")
-                if kind:
-                    hint = completion_item_kind_names[kind]
-        elif settings.completion_hint_type == "detail":
-            hint = item.get("detail")
-        elif settings.completion_hint_type == "kind":
-            kind = item.get("kind")
-            if kind:
-                hint = completion_item_kind_names.get(kind)
-        # label is an alternative for insertText if neither textEdit nor insertText is provided
-        replacement = self.text_edit_text(item) or item.get("insertText") or trigger
-        if len(replacement) > 0 and replacement[0] == '$':  # sublime needs leading '$' escaped.
-            replacement = '\\$' + replacement[1:]
-        # only return trigger with a hint if available
-        return "\t  ".join((trigger, hint)) if hint else trigger, replacement
-
-    def text_edit_text(self, item) -> 'Optional[str]':
-        text_edit = item.get("textEdit")
-        if text_edit:
-            edit_range, edit_text = text_edit.get("range"), text_edit.get("newText")
-            if edit_range and edit_text:
-                edit_range = Range.from_lsp(edit_range)
-                last_start = self.last_location - len(self.last_prefix)
-                last_row, last_col = self.view.rowcol(last_start)
-                if last_row == edit_range.start.row == edit_range.end.row and edit_range.start.col <= last_col:
-                    # sublime does not support explicit replacement with completion
-                    # at given range, but we try to trim the textEdit range and text
-                    # to the start location of the completion
-                    return edit_text[last_col - edit_range.start.col:]
-        return None
-
     def handle_response(self, response: 'Optional[Dict]'):
 
         if self.state == CompletionState.REQUESTING:
-            items = []  # type: List[Dict]
-            if isinstance(response, dict):
-                items = response["items"] or []
-            elif isinstance(response, list):
-                items = response
-            items = sorted(items, key=lambda item: item.get("sortText") or item["label"])
-            self.completions = list(self.format_completion(item) for item in items)
+            last_start = self.last_location - len(self.last_prefix)
+            last_row, last_col = self.view.rowcol(last_start)
+            self.completions = parse_completion_response(response, last_col, settings)
 
             # if insert_best_completion was just ran, undo it before presenting new completions.
             prev_char = self.view.substr(self.view.sel()[0].begin() - 1)
